Direct matches by URL path work again
This commit fixes a recent regression as a result of 5225a5741197a7e2123437709c2468e4537b0469
with the determination of non-pattern vs pattern URLs. That in turn affects the ability to perform
direct matches by URL path.

There is also a fix in PathPattern to recognize "catch-all" patterns as pattern syntax.

See gh-24945